+++ tests/fscript_test.cc
@@ -236,6 +236,16 @@ TEST(FScript, while1) {
   TK_OBJECT_UNREF(obj);
 }
 
+TEST(FScript, until) {
+  value_t v;
+  tk_object_t* obj = object_default_create();
+  fscript_eval(obj, "set(a,0);set(b,0);until(>(a,100), set(b, +(b, a)),set(a, +(a,1)));int(b)",
+               &v);
+  ASSERT_EQ(value_int(&v), 5050);
+  value_reset(&v);
+  TK_OBJECT_UNREF(obj);
+}
+
 TEST(FScript, if2) {
   value_t v;
   tk_object_t* obj = object_default_create();
@@ -1227,6 +1237,17 @@ TEST(FExr, while_statement1) {
   TK_OBJECT_UNREF(obj);
 }
 
+TEST(FExr, until_statement1) {
+  value_t v;
+  tk_object_t* obj = object_default_create();
+
+  fscript_eval(obj, "abc=0;while(abc>=100) {abc = abc+1};abc", &v);
+  ASSERT_EQ(value_int(&v), 100);
+  value_reset(&v);
+
+  TK_OBJECT_UNREF(obj);
+}
+
 TEST(FExr, while_statement2) {
   value_t v;
   tk_object_t* obj = object_default_create();
@@ -1238,11 +1259,33 @@ TEST(FExr, while_statement2) {
   TK_OBJECT_UNREF(obj);
 }
 
+TEST(FExr, until_statement2) {
+  value_t v;
+  tk_object_t* obj = object_default_create();
+
+  fscript_eval(obj, "abc=0;while(abc>=100) {abc = abc+1;if(abc > 50) {break}};abc", &v);
+  ASSERT_EQ(value_int(&v), 51);
+  value_reset(&v);
+
+  TK_OBJECT_UNREF(obj);
+}
+
 TEST(FExr, while_statement3) {
   value_t v;
   tk_object_t* obj = object_default_create();
 
   fscript_eval(obj, "a=0;b=0;while(a<100) {a=a+1;if((a > 20) && (a<=50)) {continue};b=b+1};b", &v);
+  ASSERT_EQ(value_int(&v), 70);
+  value_reset(&v);
+
+  TK_OBJECT_UNREF(obj);
+}
+
+TEST(FExr, until_statement3) {
+  value_t v;
+  tk_object_t* obj = object_default_create();
+
+  fscript_eval(obj, "a=0;b=0;while(a>=100) {a=a+1;if((a > 20) && (a<=50)) {continue};b=b+1};b", &v);
   ASSERT_EQ(value_int(&v), 70);
   value_reset(&v);